phase 3 of issue #159
